% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderLookup.R
\name{match_orders}
\alias{match_orders}
\title{Find equivalent transaction in TSMC POS system for every Order generated
through InstaCart.}
\usage{
match_orders(
  df,
  start.date = NULL,
  end.date = NULL,
  r1010.user.name = NULL,
  r1010.password = NULL
)
}
\arguments{
\item{df}{Can be either a dataframe exported from InstaCart containing all original
columns or string for a CSV file exported from InstaCart.}

\item{start.date}{Lower date limit of the search on 1010.}

\item{end.date}{Upper date limit of the search on 1010.}

\item{r1010.user.name}{String for 1010's user name. If NULL, input will be prompted in the Console.}

\item{r1010.password}{String for 1010's password. If NULL, input will be prompted in the Console.}
}
\value{
returns a dataframe containing InstaCart Order Id's and their equivalent
Transaction ID in our TMSC Pos System.
}
\description{
Find equivalent transaction in TSMC POS system for every Order generated
through InstaCart.
}
\examples{

}
